--- client/src/components/common/Calender/MainCalender.css
@@ -1,180 +1,0 @@
-/*캘린더 css*/
-
-body {
-  .react-datepicker {
-    font-family: 'Noto Sans KR', serif;
-    border: none;
-    border-radius: 0;
-    color: white;
-
-    .react-datepicker__header {
-      background-color: #232120;
-      color: white;
-      border-bottom: none;
-      border-radius: 5px;
-
-      .react-datepicker__day-names {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        background-color: #232120 !important;
-
-        .react-datepicker__day-name {
-          color: white;
-        }
-      }
-    }
-
-    .react-datepicker__triangle {
-      display: none;
-    }
-  }
-
-  .react-datepicker__month-container {
-    width: 100%;
-    height: 100%;
-  }
-  .react-datepicker__day-name {
-    font-family: Pretendard;
-    font-style: normal;
-    font-weight: 500;
-
-    width: 35px;
-    font-size: medium !important;
-    color: rgba(255, 255, 255, 0.7) !important;
-  }
-
-  .react-datepicker__day {
-    font-family: Pretendard;
-    font-style: normal;
-    font-weight: 400;
-    color: rgb(255, 255, 255) !important;
-    font-size: large !important;
-    width: 35px;
-    &--in-range:not(&--in-selecting-range) {
-      background-color: transparent;
-      color: white !important;
-    }
-  }
-  .react-datepicker__month {
-    background-color: #232120 !important;
-  }
-  .react-datepicker__month-container {
-    background-color: #232120 !important;
-  }
-
-  .react-datepicker__day--in-selecting-range,
-  .react-datepicker__day--in-range {
-    background-color: rgba(86, 201, 204, 0.4) !important;
-    color: black;
-  }
-
-  .react-datepicker__day--selecting-range-start,
-  .react-datepicker__day--selecting-range-end {
-    background-color: #56c9cc !important;
-    color: black !important;
-  }
-
-  .react-datepicker__day--disabled {
-    color: #5f5f5f !important;
-  }
-  .react-datepicker__day--outside-month {
-    color: #5f5f5f !important;
-  }
-  .react-datepicker__day--selected {
-    color: black !important;
-    background-color: #56c9cc !important;
-  }
-  .react-datepicker__day--keyboard-selected {
-    background-color: transparent;
-  }
-
-  .customHeaderContainer {
-    display: flex;
-    justify-content: space-between !important;
-    align-items: center !important;
-    background-color: #232120;
-    padding-left: 8%;
-    width: 100% !important;
-    height: 100%;
-
-    .month {
-      color: 'transparent' !important;
-      border: none;
-      font-size: 18px !important;
-      font-weight: 600;
-      margin-left: 5px;
-      cursor: pointer;
-      background: transparent;
-      color: white;
-
-      font-family: Pretendard;
-      font-style: normal;
-      font-weight: 500;
-    }
-
-    .year {
-      font-family: Pretendard;
-      font-style: normal;
-      font-weight: 500;
-
-      background-color: 'transparent' !important;
-      color: 'white' !important;
-      border: 'none' !important;
-
-      padding-right: 5px;
-      font-weight: 500;
-      cursor: pointer;
-      background: transparent;
-      border: transparent;
-      color: white;
-      max-height: '200px';
-      font-size: 18px !important;
-    }
-    .monthButton {
-      width: 34px;
-      height: 34px;
-      padding: 5px;
-      border-radius: 50%;
-      color: white !important;
-    }
-    .monthIconContainer {
-      width: 130px;
-      display: flex;
-      justify-content: space-evenly;
-      margin-left: 30px !important;
-    }
-  }
-  .calenderWrapper {
-    background-color: transparent !important;
-  }
-  .customHeaderContainer {
-    height: 50px;
-  }
-  .selectContainer {
-    display: flex;
-    align-items: center;
-  }
-  .custom-react-datepicker__wrapper {
-    position: relative;
-  }
-  .hbMOSB {
-    background-color: #232120;
-  }
-  .react-datepicker__input-container {
-    > input {
-      background-color: #413f3f !important;
-      border-radius: 5px;
-      width: 300px;
-      height: 50px;
-    }
-  }
-  .custom-react-datepicker__wrapper {
-    width: 300px;
-  }
-  .react-datepicker__month-container {
-    border-radius: 10px !important;
-    background-color: #232120 !important
-    ;
-  }
-}
