+++ client/src/components/common/Calender/calender.css
@@ -31,8 +31,6 @@ body {
   }
 
   .react-datepicker__month-container {
-    /* width: 312px;
-    height: 301px; */
     width: 100%;
     height: 100%;
   }
@@ -133,9 +131,6 @@ body {
       max-height: '200px';
       font-size: 18px !important;
     }
-    /* .selectContainer {
-      height: 100px !important;
-    } */
     .monthButton {
       width: 34px;
       height: 34px;
@@ -153,9 +148,6 @@ body {
   .calenderWrapper {
     background-color: transparent !important;
   }
-  .customHeaderContainer {
-    height: 50px;
-  }
   .selectContainer {
     display: flex;
     align-items: center;
@@ -165,11 +157,6 @@ body {
   }
   .hbMOSB {
     background-color: #232120;
-  }
-  .react-datepicker__input-container {
-    > input {
-      background-color: #413f3f !important;
-    }
   }
   .custom-react-datepicker__wrapper {
     width: 312px;
