+++ client/src/pages/main/MainBody/MainBody.tsx
@@ -10,7 +10,7 @@ import KaKaoMap from '@components/Main/PostMap';
 import { useDetailModalStore } from '@store/modal';
 import { useSearchParams } from 'react-router-dom';
 import Pagination from '@components/common/Pagination';
-import MainCalender from '@/components/common/Calender/MainCalender';
+import Calender from '@/components/common/Calender/Calender';
 
 const MainBody = (props: MainBodyPropType) => {
   const [state, setState] = useState<{
@@ -83,7 +83,7 @@ const MainBody = (props: MainBodyPropType) => {
           labelName={'태그'}
           ListItem={tagList}
         />
-        <MainCalender />
+        <Calender isMain={true} />
         <MainSearchBox placeholder={'제목'} setState={setSearchState} />
       </S.FilterGroup>
       {!selectState ? (
